refactor(web): move turn-store, turn-coordinator, turn-selectors to domains/chat/
Per CONVENTIONS.md, code used by one feature lives inside that feature.
All three modules are consumed exclusively by domains/chat/ (24 imports
across 15 files, zero cross-domain consumers). Moving them from the
top-level stores/ directory into domains/chat/ corrects this.

Pure file moves + import path updates. Zero logic changes.

Also updates STATE_MANAGEMENT.md and CONVENTIONS.md to reflect the new
locations.
