+++ apps/web/docs/STATE_MANAGEMENT.md
@@ -178,14 +178,14 @@ References:
 - [web.dev — Sign-out best practices](https://web.dev/articles/sign-out-best-practices)
 - [React — Preserving and Resetting State](https://react.dev/learn/preserving-and-resetting-state)
 
-## Turn state lives in `stores/turn-store.ts`
+## Turn state lives in `domains/chat/turn-store.ts`
 
 Turn lifecycle (sending, thinking, streaming, idle, errored), queue
 depth, active tool-call count, and current turn identity are managed
-by the turn store at `src/stores/turn-store.ts`. It sits at the top
-level because the same store is read by chat handlers, send-message
-flow, error handlers, and reconciliation — per the
-"two-or-more domains → top-level" rule above.
+by the turn store at `src/domains/chat/turn-store.ts`. All consumers
+are within the chat domain (stream handlers, send-message flow, error
+handlers, reconciliation) so the store is colocated per the
+"single-domain → inside that domain" rule.
 
 Use `useTurnStore(selector)` in React components and
 `useTurnStore.getState()` in non-React code (stream handlers,
@@ -209,7 +209,7 @@ event.
 Production callers do **not** call `turnStore.completeTurn()` /
 `cancelGeneration()` / `onStreamError()` / `onSessionError()` /
 `onPollReconciled()` directly. They call `endTurn` from
-[`stores/turn-coordinator.ts`](../src/stores/turn-coordinator.ts) — a
+[`domains/chat/turn-coordinator.ts`](../src/domains/chat/turn-coordinator.ts) — a
 single atomic two-store transition that takes the `conversationId` and
 a terminal `reason`. This prevents the "forget to clear the processing
 key" class of bug from re-appearing in every new terminal-event path.
