+++ apps/web/src/domains/chat/components/chat-composer/chat-composer.test.tsx
@@ -14,7 +14,7 @@ import { createRef } from "react";
 import { cleanup, render } from "@testing-library/react";
 
 import type { ChatAttachment } from "@/domains/chat/components/chat-attachments/use-chat-attachments";
-import { INITIAL_TURN_STATE, type TurnState, useTurnStore } from "@/stores/turn-store";
+import { INITIAL_TURN_STATE, type TurnState, useTurnStore } from "@/domains/chat/turn-store";
 
 import { ChatComposer, computeGhostSuffix, shouldSubmitOnEnter } from "@/domains/chat/components/chat-composer/chat-composer";
