+++ apps/web/src/domains/chat/hooks/use-message-queue.ts
@@ -19,7 +19,7 @@ import {
 
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile";
 import { clearQueueStatus } from "@/domains/chat/hooks/stream-message-updaters";
-import { useTurnStore } from "@/stores/turn-store";
+import { useTurnStore } from "@/domains/chat/turn-store";
 import { deleteQueuedMessage, steerToMessage } from "@/domains/chat/api/messages";
 
 // ---------------------------------------------------------------------------
