+++ apps/web/src/domains/chat/hooks/use-stream-event-handler.ts
@@ -12,8 +12,8 @@ import { useConversationStore } from "@/stores/conversation-store";
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator";
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile";
 import { tailIsStreamingAssistant } from "@/domains/chat/hooks/stream-message-updaters";
-import { useTurnStore } from "@/stores/turn-store";
-import { endTurn } from "@/stores/turn-coordinator";
+import { useTurnStore } from "@/domains/chat/turn-store";
+import { endTurn } from "@/domains/chat/turn-coordinator";
 import { useViewerStore } from "@/stores/viewer-store";
 import type { DiskPressureStatusEventPayload } from "@/assistant/use-disk-pressure-monitor";
 import {
