+++ apps/web/src/domains/chat/utils/stream-handlers/types.ts
@@ -6,8 +6,8 @@ import type {
 import type { QueryClient } from "@tanstack/react-query";
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator";
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile";
-import type { TurnActions, TurnState } from "@/stores/turn-store";
-import type { EndTurnArgs } from "@/stores/turn-coordinator";
+import type { TurnActions, TurnState } from "@/domains/chat/turn-store";
+import type { EndTurnArgs } from "@/domains/chat/turn-coordinator";
 import type { DiskPressureStatusEventPayload } from "@/assistant/use-disk-pressure-monitor";
 import type { ChatError, PendingQuestionState } from "@/domains/chat/types";
 import type { ChatEventStream } from "@/domains/chat/api/stream";
